Add files via upload
The purpose of the study was to investigate perceptions of staff about services rendered by JUSAG in the Ashanti Region. The target population was all JUSAG members in the Ashanti Region. However, the accessible population was all permanent staff of JUSAG who have been members of the association for more than one year. 125 respondents were randomly selected. With reliability coefficients of 0.969. Descriptive statistical tool was used for the data analysis. Findings revealed most JUSAG members in the Ashanti Region are satisfied with the reputation of the association and they consider not leaving the association, even if close friends recommend another association to them. majority are however not satisfied with how welfare issues are dealt with by the executives. Moreover, although members have confidence in the leadership, they express great dissatisfaction in responsiveness, and reliability level of the executives. And finally, members lament unsatisfactory financial management practices of JUSAG leadership, resulting from no annual financial statements, excessive expenditure, and prevailing lack of transparency in the financial records.